# Code split routes, reduce the initial bundle size + loading times 