--- a/packages/build/.babelrc.js
+++ b/packages/build/.babelrc.js
@@ -53,11 +53,17 @@ module.exports = {
         ['babel-plugin-styled-components', { displayName: true, ssr: false }],
       ],
     },
+    production: {
+      plugins: [
+        ...plugins,
+        ['babel-plugin-styled-components', { displayName: false, ssr: false }],
+        'babel-plugin-lodash',
+      ],
+    },
   },
   presets: makePresets(),
   plugins: [
     ...plugins,
     ['babel-plugin-styled-components', { displayName: false, ssr: false }],
-    'babel-plugin-lodash',
   ],
 };