--- a/packages/build/.babelrc.js
+++ b/packages/build/.babelrc.js
@@ -58,5 +58,6 @@ module.exports = {
   plugins: [
     ...plugins,
     ['babel-plugin-styled-components', { displayName: false, ssr: false }],
+    'babel-plugin-lodash',
   ],
 };