--- a/packages/build/webpack/webpack.base.js
+++ b/packages/build/webpack/webpack.base.js
@@ -121,31 +121,6 @@ const configFactory = {
 /** @return {import('webpack').webpack.Configuration} */
 function createDefaultConfig() {
   return {
-    optimization: {
-      splitChunks: {
-        cacheGroups: {
-          // Vendor chunk creates a chunk file that contains files coming from import statements
-          // from node_modules. The 'initial' flag directs this group to add modules to this chunk
-          //  that were imported inside only from sync chunks.
-          defaultVendors: {
-            chunks: 'initial',
-            name: 'vendor',
-            test: /([\\/]node_modules[\\/])/,
-            // Priority states that if a module falls under many cacheGroups, then
-            // the module will be part of a chunk with a higher priority.
-          },
-          // Common chunk creates a chunk file that contains modules that were shared between
-          // at least 2 (or more) async chunks. The 'async' flag directs this group to add modules
-          // to this chunk that were specifically imported inside async chunks (dynamic imports).
-          common: {
-            chunks: 'async',
-            minChunks: 2,
-            test: /([\\/]node_modules[\\/])/,
-          },
-        },
-      },
-    },
-
     entry: {
       app: ['./src/boot'],
     },