--- a/packages/build/webpack/webpack.base.js
+++ b/packages/build/webpack/webpack.base.js
@@ -16,6 +16,7 @@ limitations under the License.
 
 const path = require('path');
 
+const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 const HtmlWebPackPlugin = require('html-webpack-plugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 const ReactRefreshPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
@@ -48,6 +49,9 @@ const configFactory = {
         template: path.join(__dirname, '/../index.ejs'),
         ...options,
       });
+    },
+    bundleAnalyzer(options) {
+      return new BundleAnalyzerPlugin({ analyzerHost: '0.0.0.0', ...options });
     },
   },
   rules: {