--- a/packages/build/webpack/webpack.dev.config.js
+++ b/packages/build/webpack/webpack.dev.config.js
@@ -35,6 +35,15 @@ function getCacheConfig() {
   return cache;
 }
 
+const plugins = [
+  configFactory.plugins.tsChecker(),
+  configFactory.plugins.reactRefresh(),
+];
+
+if (process.env.WEBPACK_ANALYZE_BUNDLE === 'true') {
+  plugins.push(configFactory.plugins.bundleAnalyzer());
+}
+
 /**
  * @type { import('webpack').webpack.Configuration }
  */
@@ -48,10 +57,7 @@ module.exports = {
   },
   devtool: process.env.WEBPACK_SOURCE_MAP || 'eval-source-map',
   mode: 'development',
-  plugins: [
-    configFactory.plugins.tsChecker(),
-    configFactory.plugins.reactRefresh(),
-  ],
+  plugins,
   module: {
     strictExportPresence: true,
     rules: [