--- a/packages/build/webpack/webpack.prod.config.js
+++ b/packages/build/webpack/webpack.prod.config.js
@@ -33,6 +33,28 @@ module.exports = {
   mode: 'production',
   optimization: {
     ...configFactory.createDefaultConfig().optimization,
+    runtimeChunk: true,
+    splitChunks: {
+      chunks: 'all',
+      minSize: 20000,
+      minRemainingSize: 0,
+      minChunks: 1,
+      maxAsyncRequests: 30,
+      maxInitialRequests: 30,
+      enforceSizeThreshold: 50000,
+      cacheGroups: {
+        defaultVendors: {
+          test: /[\\/]node_modules[\\/]/,
+          priority: -10,
+          reuseExistingChunk: true,
+        },
+        default: {
+          minChunks: 2,
+          priority: -20,
+          reuseExistingChunk: true,
+        },
+      },
+    },
     minimize: true,
   },
   module: {