--- a/packages/teleport/src/Discover/index.tsx
+++ b/packages/teleport/src/Discover/index.tsx
@@ -14,4 +14,5 @@
  * limitations under the License.
  */
 
-export { Discover } from './Discover';
+// export as default for use with React.lazy
+export { Discover as default } from './Discover';