--- a/packages/teleport/src/features.ts
+++ b/packages/teleport/src/features.ts
@@ -14,27 +14,30 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+import React from 'react';
+
 import * as Icons from 'design/Icon';
 
 import Ctx from 'teleport/teleportContext';
 import cfg from 'teleport/config';
 
 import { Feature } from './types';
-import Audit from './Audit';
-import Nodes from './Nodes';
-import Sessions from './Sessions';
-import Account from './Account';
-import Applications from './Apps';
-import Kubes from './Kubes';
-import Support from './Support';
-import Clusters from './Clusters';
-import Trust from './TrustedClusters';
-import Users from './Users';
-import Roles from './Roles';
-import Recordings from './Recordings';
-import AuthConnectors from './AuthConnectors';
-import Databases from './Databases';
-import Desktops from './Desktops';
+
+const Audit = React.lazy(() => import('./Audit'));
+const Nodes = React.lazy(() => import('./Nodes'));
+const Sessions = React.lazy(() => import('./Sessions'));
+const Account = React.lazy(() => import('./Account'));
+const Applications = React.lazy(() => import('./Apps'));
+const Kubes = React.lazy(() => import('./Kubes'));
+const Support = React.lazy(() => import('./Support'));
+const Clusters = React.lazy(() => import('./Clusters'));
+const Trust = React.lazy(() => import('./TrustedClusters'));
+const Users = React.lazy(() => import('./Users'));
+const Roles = React.lazy(() => import('./Roles'));
+const Recordings = React.lazy(() => import('./Recordings'));
+const AuthConnectors = React.lazy(() => import('./AuthConnectors'));
+const Databases = React.lazy(() => import('./Databases'));
+const Desktops = React.lazy(() => import('./Desktops'));
 
 export class FeatureClusters extends Feature {
   topNavTitle = 'Clusters';