This adds the ability to run Webpack's bundle analyzer via setting the environment variable `WEBPACK_ANALYZE_BUNDLE` to `yes`. This runs a separate process with an interactive tree map that shows exactly what files are being included in what bundle that Webpack outputs.

This also code splits out all our routes and features, so the user doesn't download unnecessary JS code on first launch, and will only download what's needed for each route.

I've also removed the maximum size for a font before it's put into the JS bundle - it was adding nearly 400kb to the main `app.js` bundle.

Here's some results on the page load, JS bundle size and lighthouse performance after each change. (note these are all production builds)

|                                    | App JS Size | Transferred | Finish | DOMContentLoaded | Loaded | Lighthouse Performance |
|:----------------------------------:|:-----------:|:-----------:|:------:|:----------------:|--------|:------------------------:|
| Initial                            |    624kb    |    2.2mb    |  382ms |       270ms      |  281ms |           88           |
| Code split private routes          |    561kb    |    2.0mb    |  336ms |       224ms      |  235ms |           90           |
| Code split private + public routes |    558kb    |    2.0mb    |  332ms |       221ms      |  231ms |           90           |
| Code split all the routes          |    520kb    |    1.9mb    |  311ms |       196ms      |  206ms |           92           |
| Code split all the routes, remove fonts from JS bundle        |    151kb    |    1.5mb    |  290ms |       179ms      |  189ms |           94           |

Before this change, we had two bundles being served to the user - `vendor` and `app`. 

![image](https://user-images.githubusercontent.com/7922109/196721258-2c337501-647c-4127-a638-a370c4b29bb7.png)

This means the user would get all the `node_modules` on initial load, and all of application code.

We also were including all of `lodash.js`, which is huge. This adds `lodash-webpack-plugin` and `babel-lodash-plugin` to create smaller bundles and only include the needed parts of lodash.

After this change, the output looks like:

![image](https://user-images.githubusercontent.com/7922109/196728376-ad82ae27-7be4-43f1-9271-586f3fafa880.png)

Which has `node_modules` distributed into the relevant bundles, that are only loaded when needed (navigating to a route for example).

This also changes Webpack's `optimization` configuration to only happen in prod builds (speeds up development cycles) and have a more fine tuned config for production builds.

We still have further code splitting to do after this, as we're still producing bundles bigger than the recommended.

![image](https://user-images.githubusercontent.com/7922109/196723740-67761c7a-dd02-46fb-9c83-2a269093ee98.png)