@hatched 2022-10-19
So it looks like pulling the fonts out and tree-shaking lodash was a significant savings 👍 Lets definitely keep that! 

However for our use case I don't think we should be splitting the route assets. Our application is typically loaded once, then cached in perpetuity until a new version is released. This means that we're not as sensitive to the very first load time vs a website where that all mighty initial load time is important. This means that to save 100ms on the initial load the user will potentially run into a UX issue where navigating between routes within the application could hang while it's pulling in the assets for that route.