@ryanclark 2022-10-19
Code splitting routes means we're not loading all our node dependencies at once. We could probably chill on the amount of code split routes if it matters, but we don't want to be loading `xterm.js` for example, unless the user needs to create an SSH session for example.

There's effort to reduce the initial latency for SSH connections in the Web UI alongside the efforts for `tsh`, so creating small minimal bundles (that, like the app and vendor bundles we have at the moment, are cached after having loaded once) is better for page load performance. Even if it's just shaving off `x` milliseconds now, it avoids the page load from getting slower and slower as the codebase grows and more external modules are pulled in.

It's more likely for an API call to fail or take too long than an asset failing to load from the proxy, so I'm not concerned about that. We could probably use something a bit better than `null` for the `Suspense` fallback, but I was just copying what we use in our current `Suspense` usage when rendering the features.