@hatched 2022-10-19
I expect that making this change will actually increase the latency for the first ssh connection in the UI as we will be shifting the loading of the assets (like xterm) to the first load of the ssh UI instead of on the first load of the app. 

The browser will have less JS to parse though which may decrease the TTI when going directly to a ssh window. I'd be interested to know that once the cache is primed if this decreases the time to `Loaded` on a direct-to-ssh pageview.