@ryanclark 2022-10-19
Here's a comparison of before/after code splitting for a new SSH session

|        | Transferred | Finish | DOMContentLoaded | Loaded | Lighthouse Performance |
|--------|:-----------:|:--------:|:------------------:|:--------:|:------------------------:|
| Before |    1.8mb    | 332ms  |       263ms      | 280ms  |           92           |
| After  |    1.3mb    | 287ms  |       172ms      | 181ms  |           94           |