@alex-kovoy 2022-10-19
(my 5cents) I would suggest to have code splitting for _public_ routes such as login and for the rest it seems nice to have.

One thing to check is `fast-refresh` if it still works as expected (like it should not _always_ trigger page reloads even for minor JSX changes and instead update on-the-fly). Previously lazy-loading used to break old `hot-reloads`.

But generally, i suggest to favor code simplicity over minor optimizations if possible.