@ryanclark 2022-10-19
Yeah hot reloading still works fine with code splitting, no full page reloads when editing a file that's been split out from the main bundle